requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=5.1.0
pdfplumber>=0.11.0
